Escape menu done
Escape menu canvas will persist via DDOL throughout all scenes. In the KBM Main Menu, it will appear when the "Quit Game" button is pressed. In a 3d scene with PlayerController, it will appear when someone presses the 'Tab' key. Movement gets frozen and cursor gets unlocked. If 'Tab' is pressed again or they click the 'Cancel' button, then the canvas disappears and the player regains control. It is currently set to 'Tab' to make testing easier in the Unity editor, but will be changed to the 'Escape' key for final product.